server-django/user/user_api/serializers.py
from rest_framework import serializers
from .models import Todo, Mahasiswa

class MahasiswaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Mahasiswa
        fields = ["username", "email", "password", "nim"]
        
class TodoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Todo
        fields = ["task", "completed", "timestamp", "updated", "mahasiswa_id"]
